pyproject.toml
[tool.poetry]
name = "matrix-space-via-updater"
version = "0.1.0"
description = "Updates the 'via' server list for rooms in a space"
authors = ["GearKite"]
license = "AGPL-3.0-or-later"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
matrix-nio = "^0.24.0"
peewee = "^3.17.5"
tomli-w = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pylint = "^3.2.3"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
